Split one-click updater code to satisfy Swift file-length budget
- UpdateController one-click flows (silent download kick, transient
  retry, restart-when-idle) move to UpdateController+OneClick.swift;
  both files now under the 500-line tracking threshold
- CmuxUpdater seams extension moves from AppDelegate.swift to
  Sources/AppDelegate+Updater.swift (with the Workspace
  foreground-command helper); AppDelegate shrinks 18 lines net
- UpdateReadyToastUITests moves to its own wired file;
  UpdatePillUITests.swift restored to its pre-branch content
- UpdateReadyToastOverlay wrapper absorbs the toast positioning so the
  ContentView mount is 2 lines; budget refreshed (+3 ContentView,
  ratchets AppDelegate and stale entries down, net total decreases)
