+++ Packages/macOS/CmuxUpdater/Sources/CmuxUpdater/UpdateController.swift
@@ -17,17 +17,17 @@ public import Foundation
 /// concrete ``UpdateLogging`` and ``UpdateActionDelegate``.
 @MainActor
 public final class UpdateController {
-    private let updater: SPUUpdater
-    private let driver: UpdateDriver
-    private let log: any UpdateLogging
-    private let clock: any UpdateClock
+    let updater: SPUUpdater
+    let driver: UpdateDriver
+    let log: any UpdateLogging
+    let clock: any UpdateClock
     private let defaults: UserDefaults
     private let fileManager: FileManager
     private let hostBundle: Bundle
     private let backgroundProbeInterval: TimeInterval
     /// Whether the running build is a cmux DEV/staging build that must never be compared against
     /// the public release appcast. See ``isDevLikeBundleIdentifier(_:)``.
-    private let isDevLikeBundle: Bool
+    let isDevLikeBundle: Bool
 
     /// Host actions the updater delegates upward (retry, relaunch prep). Forwarded to the driver.
     public weak var actionDelegate: (any UpdateActionDelegate)? {
@@ -42,24 +42,26 @@ public final class UpdateController {
     private var noUpdateDismissTask: Task<Void, Never>?
     private var backgroundProbeTask: Task<Void, Never>?
     private var recheckTask: Task<Void, Never>?
-    private var silentDownloadKickTask: Task<Void, Never>?
+    // One-click auto-update state (internal, not private: the flows that use it live in
+    // UpdateController+OneClick.swift).
+    var silentDownloadKickTask: Task<Void, Never>?
     /// Set when a background probe finds an update mid-session; consumed when Sparkle reports
     /// the session finished (`onUpdateSessionFinished`), which is when the silent download can
     /// actually start.
-    private var pendingSilentDownloadKick = false
+    var pendingSilentDownloadKick = false
     /// The version the silent-download kick already ran for. The kicked session's own
     /// `didFindValidUpdate` re-reports the same version; without this the kick would re-arm on
     /// every failed download and hammer the feed in a loop.
-    private var silentDownloadKickedVersion: String?
-    private var backgroundRetryTask: Task<Void, Never>?
-    private var restartWhenIdleTask: Task<Void, Never>?
+    var silentDownloadKickedVersion: String?
+    var backgroundRetryTask: Task<Void, Never>?
+    var restartWhenIdleTask: Task<Void, Never>?
     /// How many silent-download retries were scheduled after transient background failures;
     /// bounded so a persistent outage falls back to Sparkle's scheduled hourly check.
-    private var backgroundRetryCount = 0
-    private let backgroundRetryLimit = 3
-    private let backgroundRetryDelay: Duration = .seconds(5 * 60)
+    var backgroundRetryCount = 0
+    let backgroundRetryLimit = 3
+    let backgroundRetryDelay: Duration = .seconds(5 * 60)
     /// How often the deferred "restart when idle" loop re-evaluates the host's idle signal.
-    private let restartWhenIdlePollInterval: Duration = .seconds(30)
+    let restartWhenIdlePollInterval: Duration = .seconds(30)
 
     // Readiness retry. Sparkle's `canCheckForUpdates` exposes no push signal usable under
     // Swift 6 strict concurrency (KVO on the @MainActor `SPUUpdater` "sends" a non-Sendable
@@ -209,101 +211,6 @@ public final class UpdateController {
             log.append("attemptUpdate installing freshly resolved update")
             model.state.confirm()
         }
-    }
-
-    // MARK: - Silent background download
-
-    /// Starts Sparkle's background session (which downloads and stages the update silently when
-    /// automatic downloads are enabled) once the session that detected the update has finished.
-    ///
-    /// Without this, the payload isn't fetched until Sparkle's next scheduled check, so the
-    /// user who clicks "install" pays the download (and any transient CDN failure, #5632) at
-    /// click time. Keyed on Sparkle's `didFinishUpdateCycle` signal (not a timer); the same
-    /// 100ms teardown beat as the re-check path lets Sparkle finish tearing the session down
-    /// before a new one starts, and the `canCheckForUpdates`/idle guards keep it from
-    /// interrupting a user-initiated flow.
-    private func startSilentDownloadIfKickPending() {
-        guard pendingSilentDownloadKick else { return }
-        pendingSilentDownloadKick = false
-        guard !isDevLikeBundle else { return }
-        guard updater.automaticallyDownloadsUpdates else { return }
-        silentDownloadKickTask?.cancel()
-        silentDownloadKickTask = Task { @MainActor [weak self] in
-            defer { self?.silentDownloadKickTask = nil }
-            guard let self else { return }
-            try? await self.clock.sleep(for: .milliseconds(100))
-            guard !Task.isCancelled else { return }
-            guard self.model.state.isIdle, self.model.overrideState == nil else { return }
-            guard self.updater.canCheckForUpdates else { return }
-            self.silentDownloadKickedVersion = self.model.detectedUpdateVersion
-            self.log.append("starting silent background download of detected update")
-            self.updater.checkForUpdatesInBackground()
-        }
-    }
-
-    /// Schedules a bounded silent retry after a transient background failure (GitHub's release
-    /// CDN intermittently 504s individual objects; Sparkle itself never retries, #5632).
-    /// Non-transient failures and exhausted retries fall back to the next scheduled check.
-    private func scheduleBackgroundRetryIfTransient(_ error: any Error) {
-        guard UpdateStateModel.isTransientNetworkError(error) else { return }
-        guard backgroundRetryCount < backgroundRetryLimit else {
-            log.append("background update retry limit reached; waiting for next scheduled check")
-            return
-        }
-        backgroundRetryCount += 1
-        log.append("scheduling background update retry \(backgroundRetryCount)/\(backgroundRetryLimit)")
-        backgroundRetryTask?.cancel()
-        backgroundRetryTask = Task { @MainActor [weak self] in
-            guard let self else { return }
-            try? await self.clock.sleep(for: self.backgroundRetryDelay)
-            guard !Task.isCancelled else { return }
-            guard self.model.state.isIdle, self.updater.canCheckForUpdates else { return }
-            self.log.append("retrying silent background update download")
-            self.updater.checkForUpdatesInBackground()
-        }
-    }
-
-    // MARK: - Restart when idle
-
-    /// Defers the staged update's restart until the host reports the user is idle
-    /// (``UpdateActionDelegate/updaterIsSafeToRestartNow()``), then completes the install.
-    ///
-    /// The poll is a genuine periodic schedule on the injected clock, cancelled when the user
-    /// restarts manually or the staged install goes away (see `handleStateChange`).
-    public func requestRestartWhenIdle() {
-        guard case .installing = model.effectiveState else {
-            log.append("restart-when-idle ignored (no staged install)")
-            return
-        }
-        model.setRestartWhenIdleArmed(true)
-        log.append("restart-when-idle armed")
-        guard case .installing = model.state else {
-            // Debug override without a real staged install: arm the UI state only.
-            return
-        }
-        restartWhenIdleTask?.cancel()
-        restartWhenIdleTask = Task { @MainActor [weak self] in
-            while !Task.isCancelled {
-                guard let self else { return }
-                try? await self.clock.sleep(for: self.restartWhenIdlePollInterval)
-                guard !Task.isCancelled else { return }
-                guard self.model.isRestartWhenIdleArmed,
-                      case .installing(let installing) = self.model.state else {
-                    self.cancelRestartWhenIdleLoop()
-                    return
-                }
-                guard self.actionDelegate?.updaterIsSafeToRestartNow() == true else { continue }
-                self.log.append("restart-when-idle firing (host reports idle)")
-                self.cancelRestartWhenIdleLoop()
-                installing.retryTerminatingApplication()
-                return
-            }
-        }
-    }
-
-    private func cancelRestartWhenIdleLoop() {
-        restartWhenIdleTask?.cancel()
-        restartWhenIdleTask = nil
     }
 
     // MARK: - "No updates" auto-dismiss
